Update import paths for Carrera enum
