+++ src/core/career/enum/career.enum.ts
@@ -1,0 +1,19 @@
+export enum Carrera {
+  INGENIERIA_INDUSTRIAL = 'INGENIERIA INDUSTRIAL',
+  INGENIERIA_INFORMATICA = 'INGENIERIA INFORMATICA',
+  INGENIERIA_CIVIL = 'INGENIERIA CIVIL',
+  INGENIERIA_TELECOMUNICACIONES = 'INGENIERIA TELECOMUNICACIONES',
+  ARQUITECTURA = 'ARQUITECTURA',
+  DERECHO = 'DERECHO',
+  PSICOLOGIA = 'PSICOLOGIA',
+  FILOSOFIA = 'FILOSOFIA',
+  TEOLOGIA = 'TEOLOGIA',
+  LETRAS = 'LETRAS',
+  EDUCACION = 'EDUCACION',
+  COMUNICACION_SOCIAL = 'COMUNICACION SOCIAL',
+  ADMINISTRACION = 'ADMINISTRACION',
+  CONTADURIA = 'CONTADURIA',
+  RELACIONES_INDUSTRIALES = 'RELACIONES INDUSTRIALES',
+  SOCIOLOGIA = 'SOCIOLOGIA',
+  ECONOMIA = 'ECONOMIA',
+}
