+++ src/core/profiles/dto/find-all-payload.interface.ts
@@ -1,4 +1,4 @@
-import { Carrera } from '../entities/profile.entity';
+import { Carrera } from 'src/core/career/enum/career.enum';
 
 export class FindAllPayload {
   page: number;
