+++ src/core/profiles/dto/update-profile.dto.ts
@@ -1,7 +1,7 @@
 import { ApiProperty, PartialType } from '@nestjs/swagger';
 import { IsEnum, IsNotEmpty, IsOptional, IsString } from 'class-validator';
 import { CreateProfileDTO } from './create-profile.dto';
-import { Carrera } from '../entities/profile.entity';
+import { Carrera } from 'src/core/career/enum/career.enum';
 
 export class UpdateProfileDto extends PartialType(CreateProfileDTO) {
   @ApiProperty({ example: 'Muy proactivo', required: true })
