+++ src/core/profiles/entities/profile.entity.ts
@@ -20,26 +20,7 @@ import {
   UserProfileData,
 } from '../dto/responses.dto';
 import { LanguageProfile } from './language-profile.entity';
-
-export enum Carrera {
-  INGENIERIA_INDUSTRIAL = 'INGENIERIA INDUSTRIAL',
-  INGENIERIA_INFORMATICA = 'INGENIERIA INFORMATICA',
-  INGENIERIA_CIVIL = 'INGENIERIA CIVIL',
-  INGENIERIA_TELECOMUNICACIONES = 'INGENIERIA TELECOMUNICACIONES',
-  ARQUITECTURA = 'ARQUITECTURA',
-  DERECHO = 'DERECHO',
-  PSICOLOGIA = 'PSICOLOGIA',
-  FILOSOFIA = 'FILOSOFIA',
-  TEOLOGIA = 'TEOLOGIA',
-  LETRAS = 'LETRAS',
-  EDUCACION = 'EDUCACION',
-  COMUNICACION_SOCIAL = 'COMUNICACION SOCIAL',
-  ADMINISTRACION = 'ADMINISTRACION',
-  CONTADURIA = 'CONTADURIA',
-  RELACIONES_INDUSTRIALES = 'RELACIONES INDUSTRIALES',
-  SOCIOLOGIA = 'SOCIOLOGIA',
-  ECONOMIA = 'ECONOMIA',
-}
+import { Carrera } from 'src/core/career/enum/career.enum';
 
 @Entity()
 export class Profile {
