+++ src/core/profiles/entities/profile.interface.ts
@@ -1,7 +1,7 @@
+import { Carrera } from 'src/core/career/enum/career.enum';
 import { ExperienceI } from 'src/core/experience/entities/experience.interface';
 import { SkillI } from 'src/core/skills/entities/skill.interface';
 import { UserI } from 'src/core/users/entities/user.interface';
-import { Carrera } from './profile.entity';
 
 export interface ProfileI {
   id?: number;
@@ -12,5 +12,5 @@ export interface ProfileI {
   countryResidence?: string;
   experience?: ExperienceI[];
   skills?: SkillI[];
-  carrera?: Carrera
+  carrera?: Carrera;
 }
